:recycle: Create a battery module.
